+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+  "cSpell.words": [
+    "apar",
+    "header",
+    "is",
+    "visable"
+  ]
+}
