+++ client/src/components/Feed.js
@@ -16,23 +16,20 @@ import styles from './../sass/components/Feed.module.scss';
 
 const Feed = (props) => {
 
-  // const [pageNumber, setPageNumber] = useState(1);
+  const [pageNumber, setPageNumber] = useState(1);
  
-  // const onScroll = async (event) => {
-  //   let element = event.target
-  //   if (element.scrollHeight - element.scrollTop === element.clientHeight) {
-  //     // const id = props.allPosts[props.allPosts.length -1]._id;
-    
-  //     setPageNumber(pageNumber + 1)
-  //     console.log(pageNumber)
-  //     await getLimitedPosts(pageNumber).then(async (response) => {
-  //       const newPosts = response.data;
-  //       props.addNewPostToAllPostStore(newPosts)
-  //     }).catch((err) => {
-  //       console.log(err)
-  //     })
-  //   }
-  // } 
+  const onScroll = async (event) => {
+    let element = event.target
+    if (element.scrollHeight - element.scrollTop === element.clientHeight) {
+      setPageNumber(pageNumber + 1)
+      await getLimitedPosts(pageNumber).then(async (response) => {
+        const newPosts = response.data;
+        props.addNewPostToAllPostStore(newPosts)
+      }).catch((err) => {
+        console.log(err)
+      })
+    }
+  } 
   
    const handleAllPosts = () => {
     if (props.allPosts === null || undefined) {
@@ -71,7 +68,7 @@ const Feed = (props) => {
   };
 
   return (
-    <div className={styles.feed} >
+    <div className={styles.feed} onScroll={onScroll} >
       <div className={styles.logFeedContainer}>
         <LogStatusButton />
         {handleAllPosts()}
