+++ client/src/pages/HomePage.js
@@ -1,7 +1,4 @@
-import React, {useState} from 'react';
-
-// API Calls
-import { getLimitedPosts } from './../api/handlePost';
+import React from 'react';
 
 // Components
 import About from './../components/About';
@@ -10,31 +7,11 @@ import LoginModule from './../components/LoginModule';
 
 // Redux
 import { connect } from "react-redux";
-import { addNewPostToAllPostStore } from './../redux/actions/postActions';
 
 // CSS
 import styles from './../sass/pages/HomePage.module.scss';
 
-
 const HomePage = (props) => {
-
-  const [pageNumber, setPageNumber] = useState(1);
-
-  const onScroll = async (event) => {
-    let element = event.target
-    if (element.scrollHeight - element.scrollTop === element.clientHeight) {
-      // const id = props.allPosts[props.allPosts.length -1]._id;
-
-      setPageNumber(pageNumber + 1)
-      console.log(pageNumber)
-      await getLimitedPosts(pageNumber).then(async (response) => {
-        const newPosts = response.data;
-        props.addNewPostToAllPostStore(newPosts)
-      }).catch((err) => {
-        console.log(err)
-      })
-    }
-  } 
 
   const handlePopUp = (displayPopUpLogin, displayPopUpSignUp) => {
     if (displayPopUpSignUp === true) {
@@ -61,17 +38,12 @@ const HomePage = (props) => {
   };
 
   return (
-    <div className={styles.mainContainer} onScroll={onScroll}>
+    <div className={styles.mainContainer}>
       {handlePopUp(props.displayPopUpLogin, props.displayPopUpSignUp)} 
     </div>
   )
 };
 
-const mapDispatchToProps = (dispatch) => {
-  return {
-    addNewPostToAllPostStore: (newPost) => dispatch(addNewPostToAllPostStore(newPost)) 
-  }
-};
 
 // Connects to redux store so don't need to import.
 function mapStateToProps(state) {
@@ -81,5 +53,5 @@ function mapStateToProps(state) {
   };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(HomePage);
+export default connect(mapStateToProps, null)(HomePage);
