+++ client/src/sass/components/About.module.scss
@@ -1,7 +1,7 @@
 @import './../main.scss';
 
 .about {
-	margin-top: 16rem;
+	margin-top: 10rem;
 
 	@include phone {
 		margin-top: 4rem;
