+++ client/src/sass/components/Feed.module.scss
@@ -4,6 +4,7 @@
   background-color: $color-white;
   border-radius: 25px;
   box-shadow: $shadow-2xl;
-  // height: 550px;
-  // overflow-y: scroll;
+  height: 550px;
+  overflow-y: scroll;
+  margin-bottom: 5rem;
 }
