+++ client/src/sass/pages/HomePage.module.scss
@@ -1,4 +1,4 @@
 .mainContainer {
-  height: 100vh;
-  overflow-y: scroll;
+  height: 110vh;
+  overflow: scroll;
 }
