+++ package.json
@@ -28,6 +28,7 @@
     "axios": "^0.19.2",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
+    "classnames": "^2.2.6",
     "concurrently": "^5.1.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
